requirements.txt
Flask==3.0.3
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
ultralytics==8.1.34
gunicorn